Create 4-10
This program allows the user to enter a file name (in this case, we were working with a file that listed many email conversations) and pull out email addresses from the lines. This program counts the number of messages from each person using a dictionary. After the data has been read, the person with the most emails is printed. The dictionary is made into a list and then sorted in reverse order.